requirements.txt
streamlit
pypdf
openai
python-dotenv
pandas